refactor(site): fold spend limit helper types
